+++ packages/utils/gFunc.js
@@ -224,15 +224,43 @@ export function getStorage(data, isSessionFirst = false) {
   }
   return null;
 }
-
-export function clearStorage(str) {
+/**
+ * 
+ * @param {*} str 需要清空的localStorage或sessionStorage, 如果不传清空所有
+ * @param {*} param1 需要排除的sessionStorage或者localStorage
+ * @example
+ * clearStorage()
+ * clearStorage('loginId')
+ * clearStorage('', {exceptSessions: ['loginId']});
+ */
+export function clearStorage(str='', { exceptSessions = [], exceptLocals = [] } = {}) {
+  let sessionList = {};
+  let localList = {};
+  if (exceptSessions.length > 0) {
+    exceptSessions.forEach((v) => {
+      sessionList[v] = $pub.getStorage(v, true);
+    });
+  }
+  if (exceptLocals.length > 0) {
+    exceptLocals.forEach((v) => {
+      localList[v] = $pub.getStorage(v);
+    });
+  }
   if (!str) {
     sessionStorage.clear();
     localStorage.clear();
-    return;
+  }else{
+    sessionStorage.removeItem(str);
+    localStorage.removeItem(str);
   }
-  sessionStorage.removeItem(str);
-  localStorage.removeItem(str);
+  for (const key in sessionList) {
+    const value = sessionList[key];
+    $pub.setStorage(key, value);
+  }
+  for (const key in localList) {
+    const value = localList[key];
+    $pub.setStorage(key, value, true);
+  }}
 }
 
 /**